02_fastapi_chat/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
google.generativeai